Rename backend files
