+++ app/lib/api/routes/workflow-controller.js
@@ -11,12 +11,12 @@
 
 const { Router } = require('express');
 const router = Router();
+const renderer = require('../../util/renderer');
 
+// TODO: implement required routes
 router.get('/', (req, res) => {
-  res.json({ '_links': {
-    'self': { method: 'GET', href: req.header('host') + '/quantme' },
-    'qrms': { method: 'GET', title: 'Get all available QRMs', href: req.header('host') + '/quantme/qrms' }
-  } });
+  renderer.send('test:action', 'quit');
+  return res.send('Workflow resource...');
 });
 
 module.exports = router;
