+++ app/lib/quantme/qrm-manager/git-handler.js
